requirements.txt
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
asyncclick==7.1.2.3
billiard==3.6.4.0
borb==2.1.5
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.4
django-autoslug==1.9.8
django-cities-light==3.9.2
django-dotenv==1.4.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
fonttools==4.38.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.48
Pillow==9.1.1
progressbar2==4.1.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyee==7.0.4
PyJWT==2.4.0
pyparsing==3.0.9
pyppeteer2==0.2.2
python-barcode==0.14.0
python-utils==3.3.3
pytz==2022.1
qrcode==7.3.1
redis==4.3.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
screeninfo==0.6.7
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
tqdm==4.64.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.9
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.14.1
